bots/morning-bot/requirements.txt
bubbletea-chat
openai
requests
python-dotenv
supabase
pytz
schedule
